Remove administrative work from the serving thread.
Previously, all work was done in a single thread. This might have introduced
latency, for instance, when processing responses from namerd.  Furthermore, in
order to add an admin server in an upcoming release, we'll need to have a
dedicated thread for administrative tasks.

This is accomplished by splitting configuration so that each runner has its own
reactor. The threads communicate via synchronization primitives like
`futures::sync::mpsc::channel` -- in this case so that namerd streams are
computed in the admin thread and received in the serving thread.