Use and_then instead of unwrap in main function
When reading the config file, use and_then instead of unwrapping.
This way, if an error is encountered, the user will not only see
the panic, but also the additional text "failed to read config".
It also makes it easier in the future to switch to Result based
error handling.