Ensure that remote streams are shutdown appropriately
Previously, when one end of the proxy connection stopped writing (so that
read() returns 0), the shutdown was not propagated to the other connection.
This is because the AsyncWrite::shutdown() api does nothing to actually
shutdown the connection.

We introduce Socket::tcp_shutdown to provide a consistent api for triggering a
TcpStream::shutdown.

This has been confirmed to resolve the issue with manual testing.