Ensure that remote streams are shutdown appropriately (#29)
Previously, when one end of the proxy connection stopped writing (so that
read() returns 0), the shutdown was not propagated to the other connection.
This is because the AsyncWrite::shutdown() api does nothing to actually
shutdown the connection.

We introduce Socket::tcp_shutdown to provide a consistent api for triggering a
TcpStream::shutdown.  ProxyStream uses both the AsyncWrite::shutdown and
Socket::tcp_shutdown APIs to signal when a write is complete.

This has been confirmed to fix the issue with manual testing.