Add a programatic way to exit linkerd-tcp (#24)
Problem:

Some profiling tools are not well suited for long running server
processes and rely on the process exiting to finish their work.

Solution:

POST /quitquitquit is an out-of-band signal for exiting the process. POST
is used to prevent accidental curling (which uses GET by default) or
web browsing from exiting the process.